create a `legi.db` submodule
